Crate oden_plugin_rs

Oden Rust Plugin Wrapper

oden_plugin_rs is a Rust wrapper for the Oden C plugin API, enabling users to write plugins in Rust that calls the C API.

Oden Plugin API Overview

Oden offers an API for creating plugins. Several functions are called by Oden during the lifetime of a plugin. These functions, together with some meta data, are registered using a macro. Examples of what plugins can be used for:

Visualize sensor data

Support control surfaces

Data analysis

Autonomous functions (based on sensor data)

Create custom 2D or 3D GUI:s

Custom visualizations

DLL and .so

Plugins compile to .dll files on Windows and .so files on Linux. Oden will find plugins by looking for .dll/.so files in the Oden installation directory and in the project specific plugin directory.

Note: The UUID that identifies a plugin needs to be unique for each plugin. Generators for UUIDs can be found freely online.

Plugin Structure

Filename: hello_world.rs

use oden_plugin_rs::{
    math::{Vec3, Vec4},
    register_plugin, GuiParams, InitParams, OdenPlugin, ShutdownParams, UpdateParams,
};

struct State {
    _dummy: bool,
}

impl OdenPlugin for State {
    fn init(_api: &InitParams) -> Self {
        State { _dummy: false }
    }

    fn shutdown(self, api: &ShutdownParams) {}

    fn update(&mut self, api: &UpdateParams) {
        api.im_draw_add_text(
            "Hello World!",
            Vec3::xyz(-1.0, 0.0, -2.0),
            Vec3::xyz(0.0, 0.3, 0.0),
            Vec3::xyz(1.0, 0.0, 0.0),
            Vec4::rgba(0.8, 0.8, 0.8, 1.0),
            0,
        );
    }

    fn gui(&mut self, api: &GuiParams) {
        api.label("Hello World!");
    }
}

Init

OdenPlugin::init is called when the plugin is initialized. This is an appropriate place to create data for storing state and/or initializing threads.

Update and GUI

The Oden Application runs a loop in which it calls OdenPlugin::update on all entities and then allows them to draw a GUI, and then does the drawing to screen. The rate of this loop is displayed in the top right corner (in Hz / fps).

OdenPlugin::update is called once per Oden iteration for all non-hidden entities. This is the proper place to do drawing or light recurring compute tasks.

Note: For heavier compute-tasks a separate thread should be spawned from OdenPlugin::init.

The OdenPlugin::gui function is called if the sidebar GUI of a plugin is visible. This is a good place to inspect and alter variables and show debug information in an easy way.

Draw

The function OdenPlugin::draw allows for using raw OpenGL commands.

Note: This is very complex and is not covered in the example plugin.

Shutdown

The OdenPlugin::shutdown call comes when the plugin is closing down. Data allocated during OdenPlugin::init must be freed here.

API Reference for OdenPlugin trait

InitParams contains the functions implemented for the init function.
UpdateParams contains the functions implemented for the update function.
GuiParams contains the functions implemented for the gui function.
DrawParams contains the functions implemented for the draw function. \

Functions In-parameters

There are some global functions available in the plugin API (accessed via the oden global variable). The global API functions are safe to call from all functions and they are also safe to call from other threads. All other API calls are available on parameters passed into each of the functions that get called from Oden (init, update, gui, shutdown, draw). It is safe to call all functions available in the api parameter given in each function call. It is not safe to call functions passed to one of the functions in any of the other functions, or to call them from another thread.

Data Management

All data needs to be managed and created in a way that allows for multiple instances of each plugin to coexist.

Important: Do not use global or static variables. It will prevent multiple plugin instances and is not thread safe.

Control and Names for Entities

Many functions in the plugin API use names to identify entities. When entities are added to the scene graph in Oden they are default created without names. To make these functions work, it is important to give them names.

With all entity related functions, it is also possible to concatenate @ and a scene name to entity in order to get/set an entity in a specific scene, e.g. Remote Streamer@Input if the entity is called Remote Streamer and the scene Input. This is useful if the plugin is running in a different scene to the entity that you want.

Note: Entities are named using the “Name” option available when selecting the entity in the scene graph.

Immediate Mode: GUI and ImDraw

Oden uses a paradigm called “immediate mode” for GUI and drawing functions (ImDraw). An immediate mode GUI is different from a traditional retained mode GUI. In retained mode GUIs a widget is registered and callbacks or state variables are set up that signals when things happen in the GUI. In contrast, an immediate mode GUI is drawn in each frame and the actions happen in the same code that draws the gui.

Dear ImGui

The GUI framework that Oden uses for the sidebar GUI is called “Dear ImGui”. Documentation is available for Dear ImGui online.

Label Uniqueness

Dear ImGui uses canonical label strings to identify widgets. These need to be unique.

Raw Recording

A plugin may be used to provide support for something that provides or accepts real time data. Writing a plugin allows visualizing sensor data in Oden, and at the same time e.g. show video from cameras. Oden provides a raw recording functionality to save exact data coming into Oden for later playback.

Raw recording is useful in many scenarios, e.g. algorithm development. To support the same kind of functionality as Oden support for raw recording, there are functions provide during update and also a separate function call available for playback.

During update, data can be sent to raw recording from the plugin. During playback the recordingPlayback function will be called and will be passed data that was sent from update when the system was live.

Hot Reloading

During development it is important to have a quick iteration time from code change to running application. During the development of a plugin Oden therefore offers hot reloading, so if the plugin .dll/.so is updated while Oden is running then the plugin will be reloaded with the new code.

Hot reloading resets the data state of the plugin (shutdown is called on the old plugin and init is called on the new). In some cases this is not desired, and for these cases Oden offers data migration functionality.

Important Note: Hot reloading is highly discouraged for production use and should only be used in development!

Copyright Notice

Copyright 2014-2024 Voysys AB. All rights reserved. Voysys is a registered trademark. All other company and product names may be trade marks of their respective companies with which they are associated.

Re-exports

pub use block_transfer::BlockTransferClient;

pub use block_transfer::BlockTransferServer;

pub use draw_api::DrawApi;

pub use graphics_api::GraphicsApi;

pub use graphics_api::InitAndGraphicsApi;

pub use gui_api::GuiApi;

pub use im_draw_api::ImDrawApi;

pub use init_and_update_api::FontDescriptor;

pub use init_and_update_api::InitAndUpdateApi;

pub use scene_api::CalibrationQuality;

pub use scene_api::GamepadAxis;

pub use scene_api::GamepadButton;

pub use scene_api::JoystickState;

pub use scene_api::KeyModifiers;

pub use scene_api::PerformanceMetrics;

pub use scene_api::PlaybackTimes;

pub use scene_api::QueryError;

pub use scene_api::SceneApi;

pub use scene_api::SceneApiExt;

pub use settings_api::SettingsApi;

pub use settings_api::SettingsExtApi;

pub use update_api::UpdateApi;

pub use log;

pub use zerocopy;

pub use glam;

pub use nalgebra;

pub use egui;

pub use strum;

Modules

async_com_channel_api

Async com channels can be used to send com channels messages from threads

block_transfer

Block transfer is a system where a BlockTransferServer can be used to serve a block of data and a BlockTransferClient can be used to retrieve the data.

draw_api

The Draw Api is a trait that has all the functions that are available in an draw context in Oden.

global

Global functions that are available on all threads

graphics_api

The Graphics Api is a trait that has all the functions that are used to work with Oden’s OpenGL context.

gui_api

The Gui Api is a trait that has all the functions that are available in an gui context in Oden.

im_draw_api

The Im Draw Api is a trait that has all the functions that are used for drawing simple shapes in Oden.

init_and_update_api

The Init and Update Api is a trait that has all the functions that are available in an both the init and the update context.

input_api

Input API

math

named_mpmc_channel

Named mpmc channels is a method to communicate with a channel api between different plugins

named_shared_data

Named shared data is a method for rust plugins compiled together with the same compiler to share data

oden_egui

scene_api

The Scene Api is a trait that has all the functions that are used to manipulate the scene in Oden.

settings_api

The Settings Api is a trait that has all the functions that are used to read and write settings data.

update_api

The Update Api is a trait that has all the functions that are available in an update context in Oden.

webview_user_message

Message passing between plugins and JavaScript in Webviews

Macros

check_for_gl_error

Check for OpenGL error and report it using tracing::error.

check_for_gl_error_even_in_release

Check for OpenGL error and report it using tracing::error.

instantiate_global_entity

This macro emits the necessary code for Oden to use the code as a global plugin.

register

This macro emits the necessary code for Oden to use the code as a plugin.

register_plugin

This macro emits the necessary code for Oden to use the code as a plugin.

Structs

CameraCalibration

CameraMetadata

CharEvent

DecodedFrame

DrawParams

DrawParams

FontRangeEntry

FramebufferResizeEvent

GamepadState

GuiParams

GuiParams

ImDrawImage

InferenceSettings

InitParams

InitParams

KeyEvent

LinkP2PStatusWrapper

MouseButtonEvent

NeuralNetworkInfo

QueryMetadata

ReprojectionResult

ScrollEvent

ShutdownParams

ShutdownParams

StreamStatistics

TensorData

TensorInput

TouchInputEvent

UpdateParams

UpdateParams

ViewportInfo

WindowFocusEvent

WindowPosEvent

WindowResizeEvent

Enums

ApplicationType

AsyncComChannelReceiveError

LinkError

LinkMode

NeuralNetworkError

RegisterType

Used by the ´register!´ macro to determine how to register the plugin with Oden

RegulatorMode

SceneParamError

TouchInputAction

WindowEvent

WindowMode

Constants

CameraStreamStatePaused

CameraStreamStatePlaying

CameraStreamStateStopped

CameraStreamStateUnknown

ColliderTypeBox

ColliderTypeCylinder

EventTypeChar

EventTypeFramebufferResize

EventTypeKey

EventTypeMouseButton

EventTypeScroll

EventTypeShutdown

EventTypeTouchInput

EventTypeWindowFocus

EventTypeWindowPos

EventTypeWindowResize

FpsCameraTypeFps

FpsCameraTypeOrbit

ImDrawTextAlignmentCentered

ImDrawTextAlignmentLeft

ImDrawTextAlignmentRight

JOYSTICK_1

JOYSTICK_2

JOYSTICK_3

JOYSTICK_4

JOYSTICK_5

JOYSTICK_6

JOYSTICK_7

JOYSTICK_8

JOYSTICK_9

JOYSTICK_10

JOYSTICK_11

JOYSTICK_12

JOYSTICK_13

JOYSTICK_14

JOYSTICK_15

JOYSTICK_16

JOYSTICK_LAST

JOYSTICK_MAX_AXES

KEY_0

KEY_1

KEY_2

KEY_3

KEY_4

KEY_5

KEY_6

KEY_7

KEY_8

KEY_9

KEY_A

KEY_ACTION_PRESS

KEY_ACTION_RELEASE

KEY_ACTION_REPEAT

KEY_APOSTROPHE

KEY_B

KEY_BACKSLASH

KEY_BACKSPACE

KEY_C

KEY_CAPS_LOCK

KEY_COMMA

KEY_D

KEY_DELETE

KEY_DOWN

KEY_E

KEY_END

KEY_ENTER

KEY_EQUAL

KEY_ESCAPE

KEY_F

KEY_F1

KEY_F2

KEY_F3

KEY_F4

KEY_F5

KEY_F6

KEY_F7

KEY_F8

KEY_F9

KEY_F10

KEY_F11

KEY_F12

KEY_F13

KEY_F14

KEY_F15

KEY_F16

KEY_F17

KEY_F18

KEY_F19

KEY_F20

KEY_F21

KEY_F22

KEY_F23

KEY_F24

KEY_F25

KEY_G

KEY_GRAVE_ACCENT

KEY_H

KEY_HOME

KEY_I

KEY_INSERT

KEY_J

KEY_K

KEY_KP_0

KEY_KP_1

KEY_KP_2

KEY_KP_3

KEY_KP_4

KEY_KP_5

KEY_KP_6

KEY_KP_7

KEY_KP_8

KEY_KP_9

KEY_KP_ADD

KEY_KP_DECIMAL

KEY_KP_DIVIDE

KEY_KP_ENTER

KEY_KP_EQUAL

KEY_KP_MULTIPLY

KEY_KP_SUBTRACT

KEY_L

KEY_LAST

KEY_LEFT

KEY_LEFT_ALT

KEY_LEFT_BRACKET

KEY_LEFT_CONTROL

KEY_LEFT_SHIFT

KEY_LEFT_SUPER

KEY_M

KEY_MENU

KEY_MINUS

KEY_N

KEY_NUM_LOCK

KEY_O

KEY_P

KEY_PAGE_DOWN

KEY_PAGE_UP

KEY_PAUSE

KEY_PERIOD

KEY_PRINT_SCREEN

KEY_Q

KEY_R

KEY_RIGHT

KEY_RIGHT_ALT

KEY_RIGHT_BRACKET

KEY_RIGHT_CONTROL

KEY_RIGHT_SHIFT

KEY_RIGHT_SUPER

KEY_S

KEY_SCROLL_LOCK

KEY_SEMICOLON

KEY_SLASH

KEY_SPACE

KEY_T

KEY_TAB

KEY_U

KEY_UNKNOWN

KEY_UP

KEY_V

KEY_W

KEY_WORLD_1

KEY_WORLD_2

KEY_X

KEY_Y

KEY_Z

LogCritical

LogDebug

LogError

LogInfo

LogNone

LogTrace

LogWarning

MOD_ALT

MOD_CONTROL

MOD_SHIFT

MOD_SUPER

MOUSE_BUTTON_1

MOUSE_BUTTON_2

MOUSE_BUTTON_3

MOUSE_BUTTON_4

MOUSE_BUTTON_5

MOUSE_BUTTON_6

MOUSE_BUTTON_7

MOUSE_BUTTON_8

MOUSE_BUTTON_LAST

MOUSE_BUTTON_LEFT

MOUSE_BUTTON_MIDDLE

MOUSE_BUTTON_RIGHT

MouseCursorArrow

MouseCursorPointingHand

MouseCursorResizeEW

MouseCursorResizeNESW

MouseCursorResizeNS

MouseCursorResizeNWSE

MouseCursorTextInput

ODEN_FAULT_HMD_DISCONNECTED

ODEN_FAULT_HMD_REMOVED_FROM_HEAD

ODEN_FAULT_OCULUS_DISCONNECTED

ODEN_TENSOR_SHAPE_MAX_DIMS

PLUGIN_MOD_ALT

PLUGIN_MOD_CONTROL

PLUGIN_MOD_SHIFT

PLUGIN_MOD_SUPER

PluginImDrawTextAlignmentCentered

PluginImDrawTextAlignmentLeft

PluginImDrawTextAlignmentRight

QueryDataResultDataIdNotFound

QueryDataResultDataIsNotNotExpectedSize

QueryDataResultInvalidParameters

QueryDataResultOk

TextureFormatBGRA

TextureFormatRGB

TextureFormatRGBA

TouchInputActionMove

TouchInputActionPress

TouchInputActionRelease

VideoFormatArgb

VideoFormatBgr

VideoFormatBgra

VideoFormatInvalid

VideoFormatMono8

VideoFormatMono16

VideoFormatNv12

VideoFormatRaw

VideoFormatRaw10

VideoFormatRaw16

VideoFormatRgb

VideoFormatRgba

VideoFormatUyvy

VideoFormatUyyvyy

VideoFormatV210

VideoFormatYuv

VideoFormatYuyv

Traits

OdenPlugin

OdenPlugin Trait

Functions

reproject_point_to_camera_coords

Converts a point in space to camera coordinate for a video in a stitched video. This is useful if the user want be able to press a pixel coordinate on screen and some operation should be done on that pixel in the camera image. Example: We have a stitched video named “My Stitched video entity” with 3 videos(Stream 1, 2 and 3). When the user presses somewhere in the stitched video we will display which pixels in the original video that the mouse was over.

Type Aliases

AsyncComChannelMessageId

AsyncComChannelVarSizeMsgId

CameraCropData

CameraStreamState

GlLoader

ImDrawTextAlignment

LinkStatsInterfaceStatistics

MouseCursor

OdenUuid

StreamStatisticsEx

UploadImageParams

VideoFrame

Unions

CameraMetadataInner