oden_plugin_rs

Module async_com_channel_api

Async com channels can be used to send com channels messages from threads

Enums

AsyncReceiveA

AsyncReceiveAB

AsyncReceiveABC

AsyncReceiveABCD

AsyncReceiveABCDE

AsyncReceiveABCDEF

AsyncReceiveABCDEFG

AsyncReceiveABCDEFGH

AsyncReceiveABCDEFGHI

AsyncReceiveABCDEFGHIJ

AsyncReceiveABCDEFGHIJK

AsyncReceiveABCDEFGHIJKL

AsyncReceiveABCDEFGHIJKLM

AsyncReceiveABCDEFGHIJKLMN

AsyncReceiveABCDEFGHIJKLMNO

AsyncReceiveABCDEFGHIJKLMNOP

Functions

async_com_channel_add_message_to_send_queue

async_com_channel_add_message_to_send_queue_raw

async_com_channel_flush_message_queue

async_com_channel_receive_var_size

Receive any of the ids messages into buffer, return the index of the message received and bytes written to buffer