oden_plugin_rs::async_com_channel_api

Enum AsyncReceiveABCDEFGHI

pub enum AsyncReceiveABCDEFGHI<A, B, C, D, E, F, G, H, I> {
    A(A),
    B(B),
    C(C),
    D(D),
    E(E),
    F(F),
    G(G),
    H(H),
    I(I),
}

Variants

A(A)

B(B)

C(C)

D(D)

E(E)

F(F)

G(G)

H(H)

I(I)

Implementations

impl<A, B, C, D, E, F, G, H, I> AsyncReceiveABCDEFGHI<A, B, C, D, E, F, G, H, I>

pub fn async_com_channel_receive( message_id_a: &str, message_id_b: &str, message_id_c: &str, message_id_d: &str, message_id_e: &str, message_id_f: &str, message_id_g: &str, message_id_h: &str, message_id_i: &str, timeout: Duration, ) -> Result<Self, AsyncComChannelReceiveError>
where A: FromBytes + AsBytes + Default, B: FromBytes + AsBytes + Default, C: FromBytes + AsBytes + Default, D: FromBytes + AsBytes + Default, E: FromBytes + AsBytes + Default, F: FromBytes + AsBytes + Default, G: FromBytes + AsBytes + Default, H: FromBytes + AsBytes + Default, I: FromBytes + AsBytes + Default,