oden_plugin_rs::gui_api

Struct MockGuiApi

pub struct MockGuiApi { /* private fields */ }

Implementations

impl MockGuiApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockGuiApi

pub fn expect_label(&mut self) -> &mut Expectation

Create an Expectation for mocking the label method

pub fn expect_button(&mut self) -> &mut Expectation

Create an Expectation for mocking the button method

pub fn expect_inactive_button(&mut self) -> &mut Expectation

Create an Expectation for mocking the inactive_button method

pub fn expect_combo(&mut self) -> &mut Expectation

Create an Expectation for mocking the combo method

pub fn expect_inactive_combo(&mut self) -> &mut Expectation

Create an Expectation for mocking the inactive_combo method

pub fn expect_slider_int(&mut self) -> &mut Expectation

Create an Expectation for mocking the slider_int method

pub fn expect_slider_float(&mut self) -> &mut Expectation

Create an Expectation for mocking the slider_float method

pub fn expect_slider_float3(&mut self) -> &mut Expectation

Create an Expectation for mocking the slider_float3 method

pub fn expect_checkbox(&mut self) -> &mut Expectation

Create an Expectation for mocking the checkbox method

pub fn expect_tree_node(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the tree_node method

pub fn expect_drag_int(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_int method

pub fn expect_drag_int2(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_int2 method

pub fn expect_drag_int3(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_int3 method

pub fn expect_drag_int4(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_int4 method

pub fn expect_drag_int5(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_int5 method

pub fn expect_drag_float(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_float method

pub fn expect_drag_float2(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_float2 method

pub fn expect_drag_float3(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_float3 method

pub fn expect_drag_float4(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_float4 method

pub fn expect_drag_float5(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_float5 method

pub fn expect_drag_vec2(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_vec2 method

pub fn expect_drag_vec3(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_vec3 method

pub fn expect_drag_vec4(&mut self) -> &mut Expectation

Create an Expectation for mocking the drag_vec4 method

pub fn expect_same_line(&mut self) -> &mut Expectation

Create an Expectation for mocking the same_line method

pub fn expect_input_text(&mut self) -> &mut Expectation

Create an Expectation for mocking the input_text method

pub fn expect_label_colored(&mut self) -> &mut Expectation

Create an Expectation for mocking the label_colored method

pub fn expect_tooltip_if_last_item_hovered(&mut self) -> &mut Expectation

Create an Expectation for mocking the tooltip_if_last_item_hovered method

pub fn expect_graph(&mut self) -> &mut Expectation

Create an Expectation for mocking the graph method

pub fn expect_indent(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the indent method

pub fn expect_checkbox_settings_backed(&mut self) -> &mut Expectation

Create an Expectation for mocking the checkbox_settings_backed method

pub fn expect_color_edit3(&mut self) -> &mut Expectation

Create an Expectation for mocking the color_edit3 method

pub fn expect_color_edit4(&mut self) -> &mut Expectation

Create an Expectation for mocking the color_edit4 method

pub fn expect_entity_picker(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_picker method

pub fn expect_joystick_picker(&mut self) -> &mut Expectation

Create an Expectation for mocking the joystick_picker method

pub fn expect_joystick_picker_settings_backed(&mut self) -> &mut Expectation

Create an Expectation for mocking the joystick_picker_settings_backed method

pub fn expect_combo_enum<T: IntoEnumIterator + VariantNames + Display + PartialEq + 'static>( &mut self, ) -> &mut Expectation<T>

Create an Expectation for mocking the combo_enum method