oden_plugin_rs::im_draw_api

Struct MockImDrawApi

pub struct MockImDrawApi { /* private fields */ }

Implementations

impl MockImDrawApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockImDrawApi

pub fn expect_im_draw_set_use_z_buffer(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_set_use_z_buffer method

pub fn expect_im_draw_set_write_z_buffer(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_set_write_z_buffer method

pub fn expect_im_draw_push_matrix(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_push_matrix method

pub fn expect_im_draw_pop_matrix(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_pop_matrix method

pub fn expect_im_draw_push_view_space(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_push_view_space method

pub fn expect_im_draw_pop_view_space(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_pop_view_space method

pub fn expect_im_draw_set_view_space_z(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_set_view_space_z method

pub fn expect_im_draw_set_tint(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_set_tint method

pub fn expect_im_draw_push_shader(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_push_shader method

pub fn expect_im_draw_pop_shader(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_pop_shader method

pub fn expect_im_draw_latch_backbuffer(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_latch_backbuffer method

pub fn expect_im_draw_add_line(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_line method

pub fn expect_im_draw_add_circle_segment(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_circle_segment method

pub fn expect_im_draw_add_circle(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_circle method

pub fn expect_im_draw_add_square(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_square method

pub fn expect_im_draw_add_rectangle(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle method

pub fn expect_im_draw_add_rectangle_rounded(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_rounded method

pub fn expect_im_draw_add_image(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_image method

pub fn expect_im_draw_add_video(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_video method

pub fn expect_im_draw_add_video_with_uvs(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_video_with_uvs method

pub fn expect_im_draw_add_text(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_text method

pub fn expect_im_draw_add_text_normal(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_text_normal method

pub fn expect_im_draw_add_text_aligned(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_text_aligned method

pub fn expect_im_draw_add_text_aligned_normal(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_text_aligned_normal method

pub fn expect_im_draw_add_torus(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_torus method

pub fn expect_im_draw_add_lines(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_lines method

pub fn expect_im_draw_add_rectangle_border(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_border method

pub fn expect_im_draw_add_rectangle_border_rounded( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_border_rounded method

pub fn expect_im_draw_add_cuboid(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_cuboid method

pub fn expect_im_draw_add_triangles(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_triangles method

pub fn expect_im_draw_add_rectangle_aligned(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_aligned method

pub fn expect_im_draw_add_rectangle_rounded_aligned( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_rounded_aligned method

pub fn expect_im_draw_add_rectangle_border_aligned( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_border_aligned method

pub fn expect_im_draw_add_rectangle_border_rounded_aligned( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_rectangle_border_rounded_aligned method

pub fn expect_im_draw_add_triangles_with_uvs(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_add_triangles_with_uvs method

pub fn expect_im_draw_calc_text_size(&mut self) -> &mut Expectation

Create an Expectation for mocking the im_draw_calc_text_size method