oden_plugin_rs

Module input_api

Input API

API for handling input from joysticks and gamepads.

Structs

Gamepad

A gamepad with named buttons and axes.

HandTrackingState

OpenXR 26-joint hand tracking state.

Enums

PowerState

Power state of the system.

Constants

HAND_FLAG_ACTIVE

Hand tracking is active and data is valid.

HAND_FLAG_LEFT

This is the left hand.

HAND_FLAG_LOCAL

Joint positions are in wrist-local coordinates (default from API).

HAND_FLAG_RIGHT

This is the right hand.

Functions

add_gamepad_mapping

Add a gamepad mapping.

gamepad

Get the state of the gamepad if exactly one is connected.

gamepad_state

Get the state of a gamepad with the given id.

get_power_info

Get the power info of the system.

hand_tracking_state

Get hand tracking state. hand: 0=left, 1=right.

is_joystick_present

Check if a joystick with the given id is present.

joystick_state

Get the state of a joystick with the given id.