oden_plugin_rs

Macro instantiate_global_entity

macro_rules! instantiate_global_entity {
    ($name:expr, $uuid:expr, $state_type:ident) => { ... };
}

This macro emits the necessary code for Oden to use the code as a global plugin.

Global plugins are different from regular plugins as global plugins are not added to the scene graph. All global plugins that Oden can find will be running regardless of which project that is loaded. For example, a global plugin can be use to change project depending on an external signal.

Global plugins are registered in the same way as regular plugins except with the instantiate_global_entity macro instead as shown in the example below.

Examples

instantiate_global_entity!(
    "Hello World from Rust",
    "312fe6ca-9ada-4224-8320-7554aa21864f",
    State
);