oden_plugin_rs::named_mpmc_channel

Trait NamedMpmcChannelExt

pub trait NamedMpmcChannelExt {
    // Required methods
    fn named_mpmc_channel_tx<T>(&self, name: &str, buffer: usize) -> Sender<T>;
    fn named_mpmc_channel_rx<T>(&self, name: &str, buffer: usize) -> Receiver<T>;
}

Extension trait for named mpmc channel methods

Required Methods

fn named_mpmc_channel_tx<T>(&self, name: &str, buffer: usize) -> Sender<T>

Create or fetch a named mpmc channel with the specified name

fn named_mpmc_channel_rx<T>(&self, name: &str, buffer: usize) -> Receiver<T>

Create or fetch a named mpmc channel with the specified name

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.