oden_plugin_rs

Module scene_api

The Scene Api is a trait that has all the functions that are used to manipulate the scene in Oden.

Modules

__mock_MockSceneApi

__mock_MockSceneApi_SceneApi

Structs

CalibrationQuality

Collider

InferenceMetric

JoystickState

KeyModifiers

MockSceneApi

PerformanceMetrics

PlaybackTimes

PrintScreen

RemoteCalibrationResult

RigidBody

RigidBodyUuid

Enums

ColliderType

GamepadAxis

Gamepad axis indices corresponding to SDL3’s SDL_GamepadAxis layout.

GamepadButton

Gamepad button indices corresponding to SDL3’s SDL_GamepadButton layout, used to index GamepadState’s buttons array.

LinkP2PStatus

QueryError

Traits

SceneApi

SceneApiExt