oden_plugin_rs::scene_api::__mock_MockSceneApi_SceneApi::__set_link_p2p_stun_server

Struct Expectation

pub struct Expectation { /* private fields */ }

Expectation type for methods that return a 'static type. This is the type returned by the expect_* methods.

Implementations

impl Expectation

pub fn return_const<MockallOutput>( &mut self, __mockall_c: MockallOutput, ) -> &mut Self
where MockallOutput: Clone + Into<Result<(), LinkError>> + Send + 'static,

Return a constant value from the Expectation

The output type must be Clone. The compiler can’t always infer the proper type to use with this method; you will usually need to specify it explicitly. i.e. return_const(42i32) instead of return_const(42).

pub fn return_const_st<MockallOutput>( &mut self, __mockall_c: MockallOutput, ) -> &mut Self
where MockallOutput: Clone + Into<Result<(), LinkError>> + 'static,

Single-threaded version of return_const. This is useful for return types that are not Send.

The output type must be Clone. The compiler can’t always infer the proper type to use with this method; you will usually need to specify it explicitly. i.e. return_const(42i32) instead of return_const(42).

It is a runtime error to call the mock method from a different thread than the one that originally called this method.

pub fn return_once<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> FnOnce(Option<&'a str>, i32, &str) -> Result<(), LinkError> + Send + 'static,

Supply an FnOnce closure that will provide the return value for this Expectation. This is useful for return types that aren’t Clone. It will be an error to call this method multiple times.

pub fn return_once_st<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> FnOnce(Option<&'a str>, i32, &str) -> Result<(), LinkError> + 'static,

Single-threaded version of return_once. This is useful for return types that are neither Send nor Clone.

It is a runtime error to call the mock method from a different thread than the one that originally called this method. It is also a runtime error to call the method more than once.

pub fn returning<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> FnMut(Option<&'a str>, i32, &str) -> Result<(), LinkError> + Send + 'static,

Supply a closure that will provide the return value for this Expectation. The method’s arguments are passed to the closure by value.

pub fn returning_st<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> FnMut(Option<&'a str>, i32, &str) -> Result<(), LinkError> + 'static,

Single-threaded version of returning. Can be used when the argument or return type isn’t Send.

It is a runtime error to call the mock method from a different thread than the one that originally called this method.

pub fn in_sequence(&mut self, __mockall_seq: &mut Sequence) -> &mut Self

Add this expectation to a Sequence.

pub fn never(&mut self) -> &mut Self

Forbid this expectation from ever being called.

pub fn new() -> Self

Create a new, default, Expectation

pub fn once(&mut self) -> &mut Self

Expect this expectation to be called exactly once. Shortcut for times(1).

pub fn times<MockallR>(&mut self, __mockall_r: MockallR) -> &mut Self
where MockallR: Into<TimesRange>,

Restrict the number of times that that this method may be called.

The argument may be:

A fixed number: .times(4)

Various types of range:

.times(5..10)

.times(..10)

.times(5..)

.times(5..=10)

.times(..=10)

The wildcard: .times(..)

pub fn with<MockallMatcher0: for<'a> Predicate<Option<&'a str>> + Send + 'static, MockallMatcher1: for<'a> Predicate<i32> + Send + 'static, MockallMatcher2: for<'a> Predicate<str> + Send + 'static>( &mut self, entity: MockallMatcher0, index: MockallMatcher1, addr: MockallMatcher2, ) -> &mut Self

Set matching crieteria for this Expectation.

The matching predicate can be anything implemening the Predicate trait. Only one matcher can be set per Expectation at a time.

pub fn withf<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> Fn(&Option<&'a str>, &i32, &str) -> bool + Send + 'static,

Set a matching function for this Expectation.

This is equivalent to calling with with a function argument, like with(predicate::function(f)).

pub fn withf_st<MockallF>(&mut self, __mockall_f: MockallF) -> &mut Self
where MockallF: for<'a> Fn(&Option<&'a str>, &i32, &str) -> bool + 'static,

Single-threaded version of withf. Can be used when the argument type isn’t Send.