oden_plugin_rs::scene_api

Struct MockSceneApi

pub struct MockSceneApi { /* private fields */ }

Implementations

impl MockSceneApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockSceneApi

pub fn expect_inhibit_mouse_look(&mut self) -> &mut Expectation

Create an Expectation for mocking the inhibit_mouse_look method

pub fn expect_horizon(&mut self) -> &mut Expectation

Create an Expectation for mocking the horizon method

pub fn expect_set_horizon(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_horizon method

pub fn expect_entity_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_name method

pub fn expect_calibrated_floor_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the calibrated_floor_height method

pub fn expect_set_stitch_distance(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_stitch_distance method

pub fn expect_camera_metadata(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_metadata method

pub fn expect_remote_streamer_metadata(&mut self) -> &mut Expectation

Create an Expectation for mocking the remote_streamer_metadata method

pub fn expect_camera_temperature(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_temperature method

pub fn expect_camera_fps(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_fps method

pub fn expect_camera_last_valid_timestamp(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_last_valid_timestamp method

pub fn expect_set_entity_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_entity_enabled method

pub fn expect_camera_current_timestamp(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_current_timestamp method

pub fn expect_set_streamer_frame_metadata(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_streamer_frame_metadata method

pub fn expect_load_project(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_project method

pub fn expect_start_streamer(&mut self) -> &mut Expectation

Create an Expectation for mocking the start_streamer method

pub fn expect_stop_streamer(&mut self) -> &mut Expectation

Create an Expectation for mocking the stop_streamer method

pub fn expect_configure_texture_streaming(&mut self) -> &mut Expectation

Create an Expectation for mocking the configure_texture_streaming method

pub fn expect_start_texture_streaming(&mut self) -> &mut Expectation

Create an Expectation for mocking the start_texture_streaming method

pub fn expect_stop_texture_streaming(&mut self) -> &mut Expectation

Create an Expectation for mocking the stop_texture_streaming method

pub fn expect_camera_stream_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_stream_state method

pub fn expect_set_camera_stream_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_camera_stream_state method

pub fn expect_all_scene_ids(&mut self) -> &mut Expectation

Create an Expectation for mocking the all_scene_ids method

pub fn expect_selected_scene_id(&mut self) -> &mut Expectation

Create an Expectation for mocking the selected_scene_id method

pub fn expect_set_selected_scene(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_selected_scene method

pub fn expect_scene_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the scene_name method

pub fn expect_raw_record_start(&mut self) -> &mut Expectation

Create an Expectation for mocking the raw_record_start method

pub fn expect_raw_record_stop(&mut self) -> &mut Expectation

Create an Expectation for mocking the raw_record_stop method

pub fn expect_set_raw_record_entity_ring_buffer_size( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_raw_record_entity_ring_buffer_size method

pub fn expect_raw_record_entity_ring_buffer_enable( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the raw_record_entity_ring_buffer_enable method

pub fn expect_raw_record_last_path(&mut self) -> &mut Expectation

Create an Expectation for mocking the raw_record_last_path method

pub fn expect_entity_world_matrix(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_world_matrix method

pub fn expect_is_joystick_present(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_joystick_present method

pub fn expect_joystick_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the joystick_state method

pub fn expect_gamepad_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the gamepad_state method

pub fn expect_joystick_state_from_settings_backed_name( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the joystick_state_from_settings_backed_name method

pub fn expect_stream_statistics(&mut self) -> &mut Expectation

Create an Expectation for mocking the stream_statistics method

pub fn expect_stream_statistics_ex(&mut self) -> &mut Expectation

Create an Expectation for mocking the stream_statistics_ex method

pub fn expect_has_entity_with_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the has_entity_with_name method

pub fn expect_streamer_bitrate_mbps(&mut self) -> &mut Expectation

Create an Expectation for mocking the streamer_bitrate_mbps method

pub fn expect_streamer_link_feedback_stats(&mut self) -> &mut Expectation

Create an Expectation for mocking the streamer_link_feedback_stats method

pub fn expect_set_streamer_max_bandwidth(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_streamer_max_bandwidth method

pub fn expect_link_count(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_count method

pub fn expect_set_link_mode(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_mode method

pub fn expect_link_mode(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_mode method

pub fn expect_set_link_bind_ip(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_bind_ip method

pub fn expect_link_bind_ip(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_bind_ip method

pub fn expect_set_link_receive_port(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_receive_port method

pub fn expect_link_receive_port(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_receive_port method

pub fn expect_set_link_destination_ip(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_destination_ip method

pub fn expect_link_destination_ip(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_destination_ip method

pub fn expect_set_link_destination_port(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_destination_port method

pub fn expect_link_destination_port(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_destination_port method

pub fn expect_set_link_p2p_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_p2p_enabled method

pub fn expect_link_p2p_status(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_p2p_status method

pub fn expect_link_p2p_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_p2p_enabled method

pub fn expect_set_link_p2p_stun_server(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_p2p_stun_server method

pub fn expect_link_p2p_stun_server(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_p2p_stun_server method

pub fn expect_all_entity_uuids(&mut self) -> &mut Expectation

Create an Expectation for mocking the all_entity_uuids method

pub fn expect_entity_type_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_type_name method

pub fn expect_basic_statistics(&mut self) -> &mut Expectation

Create an Expectation for mocking the basic_statistics method

pub fn expect_remote_streamer_statistics(&mut self) -> &mut Expectation

Create an Expectation for mocking the remote_streamer_statistics method

pub fn expect_alignment_id(&mut self) -> &mut Expectation

Create an Expectation for mocking the alignment_id method

pub fn expect_video_placement(&mut self) -> &mut Expectation

Create an Expectation for mocking the video_placement method

pub fn expect_set_link_encryption_private_key_path( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_link_encryption_private_key_path method

pub fn expect_link_encryption_private_key_path(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_encryption_private_key_path method

pub fn expect_add_link_encryption_allowed_public_key( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the add_link_encryption_allowed_public_key method

pub fn expect_remove_link_encryption_allowed_public_key( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the remove_link_encryption_allowed_public_key method

pub fn expect_link_encryption_allowed_public_keys_count( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the link_encryption_allowed_public_keys_count method

pub fn expect_link_encryption_allowed_public_key(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_encryption_allowed_public_key method

pub fn expect_set_link_encryption_internal_source_ip( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_link_encryption_internal_source_ip method

pub fn expect_link_encryption_internal_source_ip(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_encryption_internal_source_ip method

pub fn expect_set_link_encryption_internal_destination_ip( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_link_encryption_internal_destination_ip method

pub fn expect_link_encryption_internal_destination_ip( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the link_encryption_internal_destination_ip method

pub fn expect_add_link(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_link method

pub fn expect_remove_link(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_link method

pub fn expect_show_hmd_controllers(&mut self) -> &mut Expectation

Create an Expectation for mocking the show_hmd_controllers method

pub fn expect_hmd_transform(&mut self) -> &mut Expectation

Create an Expectation for mocking the hmd_transform method

pub fn expect_set_focus_region_center_pixel_position( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_focus_region_center_pixel_position method

pub fn expect_entity_video_size(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_video_size method

pub fn expect_set_regulator_mode(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_regulator_mode method

pub fn expect_regulator_mode(&mut self) -> &mut Expectation

Create an Expectation for mocking the regulator_mode method

pub fn expect_set_gui_visible(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_gui_visible method

pub fn expect_is_key_down(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_key_down method

pub fn expect_is_key_pressed_since_last_swap(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_key_pressed_since_last_swap method

pub fn expect_keys_pressed_since_last_swap(&mut self) -> &mut Expectation

Create an Expectation for mocking the keys_pressed_since_last_swap method

pub fn expect_window_events(&mut self) -> &mut Expectation

Create an Expectation for mocking the window_events method

pub fn expect_raw_record_is_running(&mut self) -> &mut Expectation

Create an Expectation for mocking the raw_record_is_running method

pub fn expect_playback_start(&mut self) -> &mut Expectation

Create an Expectation for mocking the playback_start method

pub fn expect_playback_pause(&mut self) -> &mut Expectation

Create an Expectation for mocking the playback_pause method

pub fn expect_set_playback_time(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_playback_time method

pub fn expect_playback_advance_time(&mut self) -> &mut Expectation

Create an Expectation for mocking the playback_advance_time method

pub fn expect_set_playback_loop(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_playback_loop method

pub fn expect_playback_is_playing(&mut self) -> &mut Expectation

Create an Expectation for mocking the playback_is_playing method

pub fn expect_playback_times(&mut self) -> &mut Expectation

Create an Expectation for mocking the playback_times method

pub fn expect_com_channel_max_message_size(&mut self) -> &mut Expectation

Create an Expectation for mocking the com_channel_max_message_size method

pub fn expect_com_channel_send_message_raw(&mut self) -> &mut Expectation

Create an Expectation for mocking the com_channel_send_message_raw method

pub fn expect_com_channel_message_count(&mut self) -> &mut Expectation

Create an Expectation for mocking the com_channel_message_count method

pub fn expect_com_channel_var_size_message(&mut self) -> &mut Expectation

Create an Expectation for mocking the com_channel_var_size_message method

pub fn expect_com_channel_last_var_size_msg(&mut self) -> &mut Expectation

Create an Expectation for mocking the com_channel_last_var_size_msg method

pub fn expect_camera_calibration(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_calibration method

pub fn expect_scene_containing_plugin_entity(&mut self) -> &mut Expectation

Create an Expectation for mocking the scene_containing_plugin_entity method

pub fn expect_save_project(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_project method

pub fn expect_project_has_unsaved_changes(&mut self) -> &mut Expectation

Create an Expectation for mocking the project_has_unsaved_changes method

pub fn expect_set_force_feedback_angle(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_force_feedback_angle method

pub fn expect_is_streamer_running(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_streamer_running method

pub fn expect_camera_last_received_frame_time(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_last_received_frame_time method

pub fn expect_set_audio_receiving_enable(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_audio_receiving_enable method

pub fn expect_set_audio_sending_enable(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_audio_sending_enable method

pub fn expect_load_scene(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_scene method

pub fn expect_remove_scene(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_scene method

pub fn expect_link_stats_bytes_transferred(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_stats_bytes_transferred method

pub fn expect_output_alignment_position(&mut self) -> &mut Expectation

Create an Expectation for mocking the output_alignment_position method

pub fn expect_set_output_alignment_position(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_output_alignment_position method

pub fn expect_output_alignment_scale(&mut self) -> &mut Expectation

Create an Expectation for mocking the output_alignment_scale method

pub fn expect_set_output_alignment_scale(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_output_alignment_scale method

pub fn expect_output_alignment_rotation(&mut self) -> &mut Expectation

Create an Expectation for mocking the output_alignment_rotation method

pub fn expect_set_output_alignment_rotation(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_output_alignment_rotation method

pub fn expect_streamer_resolution(&mut self) -> &mut Expectation

Create an Expectation for mocking the streamer_resolution method

pub fn expect_set_streamer_resolution(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_streamer_resolution method

pub fn expect_set_sync_camera(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_sync_camera method

pub fn expect_decoder_stats(&mut self) -> &mut Expectation

Create an Expectation for mocking the decoder_stats method

pub fn expect_is_over_gui(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_over_gui method

pub fn expect_close_project(&mut self) -> &mut Expectation

Create an Expectation for mocking the close_project method

pub fn expect_request_print_screen(&mut self) -> &mut Expectation

Create an Expectation for mocking the request_print_screen method

pub fn expect_print_screen(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the print_screen method

pub fn expect_window_mode(&mut self) -> &mut Expectation

Create an Expectation for mocking the window_mode method

pub fn expect_link_index(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_index method

pub fn expect_link_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_uuid method

pub fn expect_link_stats_interface(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_stats_interface method

pub fn expect_set_bandwidth_regulator_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_bandwidth_regulator_state method

pub fn expect_bandwidth_regulator_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the bandwidth_regulator_state method

pub fn expect_scene_to_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the scene_to_string method

pub fn expect_load_scene_from_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the load_scene_from_string method

pub fn expect_scene_entity_uuids(&mut self) -> &mut Expectation

Create an Expectation for mocking the scene_entity_uuids method

pub fn expect_flexbox_layout(&mut self) -> &mut Expectation

Create an Expectation for mocking the flexbox_layout method

pub fn expect_set_flexbox_padding(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_flexbox_padding method

pub fn expect_target_bitrate(&mut self) -> &mut Expectation

Create an Expectation for mocking the target_bitrate method

pub fn expect_entity_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_enabled method

pub fn expect_output_alignment_id(&mut self) -> &mut Expectation

Create an Expectation for mocking the output_alignment_id method

pub fn expect_clone_source_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the clone_source_uuid method

pub fn expect_set_clone_stream_source(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_clone_stream_source method

pub fn expect_rigid_bodies(&mut self) -> &mut Expectation

Create an Expectation for mocking the rigid_bodies method

pub fn expect_rigid_bodies_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the rigid_bodies_uuid method

pub fn expect_set_camera_calibration(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_camera_calibration method

pub fn expect_set_camera_calibration_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_camera_calibration_string method

pub fn expect_dewarping_params(&mut self) -> &mut Expectation

Create an Expectation for mocking the dewarping_params method

pub fn expect_set_dewarping_params(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_dewarping_params method

pub fn expect_gui_visible(&mut self) -> &mut Expectation

Create an Expectation for mocking the gui_visible method

pub fn expect_set_virtual_camera_position(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_virtual_camera_position method

pub fn expect_set_virtual_camera_rotation(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_virtual_camera_rotation method

pub fn expect_set_virtual_camera_fov(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_virtual_camera_fov method

pub fn expect_set_virtual_camera_resolution(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_virtual_camera_resolution method

pub fn expect_virtual_camera_resolution(&mut self) -> &mut Expectation

Create an Expectation for mocking the virtual_camera_resolution method

pub fn expect_calibration_quality(&mut self) -> &mut Expectation

Create an Expectation for mocking the calibration_quality method

pub fn expect_add_entity(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_entity method

pub fn expect_remove_entity(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_entity method

pub fn expect_poll_remote_calibration(&mut self) -> &mut Expectation

Create an Expectation for mocking the poll_remote_calibration method

pub fn expect_camera_crop(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_crop method

pub fn expect_set_camera_crop(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_camera_crop method

pub fn expect_camera_hard_crop(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_hard_crop method

pub fn expect_set_camera_hard_crop(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_camera_hard_crop method

pub fn expect_drop_detector_timeout(&mut self) -> &mut Expectation

Create an Expectation for mocking the drop_detector_timeout method

pub fn expect_set_drop_detector_timeout(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_drop_detector_timeout method

pub fn expect_show_no_signal_screen(&mut self) -> &mut Expectation

Create an Expectation for mocking the show_no_signal_screen method

pub fn expect_set_show_no_signal_screen(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_show_no_signal_screen method

pub fn expect_max_reorder_frames(&mut self) -> &mut Expectation

Create an Expectation for mocking the max_reorder_frames method

pub fn expect_set_max_reorder_frames(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_max_reorder_frames method

pub fn expect_regulator_max_packet_size(&mut self) -> &mut Expectation

Create an Expectation for mocking the regulator_max_packet_size method

pub fn expect_set_regulator_max_packet_size(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_regulator_max_packet_size method

pub fn expect_entity_scale(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_scale method

pub fn expect_set_entity_scale(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_entity_scale method

pub fn expect_raw_record_folder(&mut self) -> &mut Expectation

Create an Expectation for mocking the raw_record_folder method

pub fn expect_set_raw_record_folder(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_raw_record_folder method

pub fn expect_virtual_camera_render_only_uuids(&mut self) -> &mut Expectation

Create an Expectation for mocking the virtual_camera_render_only_uuids method

pub fn expect_set_external_stitch_geometry(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_external_stitch_geometry method

pub fn expect_ray_to_video_intersect(&mut self) -> &mut Expectation

Create an Expectation for mocking the ray_to_video_intersect method

pub fn expect_parent_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the parent_uuid method

pub fn expect_ancestor_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the ancestor_uuid method

pub fn expect_entity_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the entity_uuid method

pub fn expect_flexbox_rotation(&mut self) -> &mut Expectation

Create an Expectation for mocking the flexbox_rotation method

pub fn expect_performance_metrics(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the performance_metrics method

pub fn expect_set_link_encryption_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_encryption_enabled method

pub fn expect_clear_link_encryption_allowed_public_keys( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the clear_link_encryption_allowed_public_keys method

pub fn expect_set_link_to_relay_link(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_to_relay_link method

pub fn expect_is_link_relay_link(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_link_relay_link method

pub fn expect_set_link_send_to_relay(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_send_to_relay method

pub fn expect_is_link_sending_to_relay(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_link_sending_to_relay method

pub fn expect_set_link_drop_outgoing_packets(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_drop_outgoing_packets method

pub fn expect_remote_streamer_video_info(&mut self) -> &mut Expectation

Create an Expectation for mocking the remote_streamer_video_info method

pub fn expect_stream_name(&mut self) -> &mut Expectation

Create an Expectation for mocking the stream_name method

pub fn expect_set_packer_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_packer_enabled method

pub fn expect_set_packer_auto_crop(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_packer_auto_crop method

pub fn expect_is_packer_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_packer_enabled method

pub fn expect_is_packer_auto_crop_enabled(&mut self) -> &mut Expectation

Create an Expectation for mocking the is_packer_auto_crop_enabled method

pub fn expect_set_background_color(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_background_color method

pub fn expect_background_color(&mut self) -> &mut Expectation

Create an Expectation for mocking the background_color method

pub fn expect_clear_background_color(&mut self) -> &mut Expectation

Create an Expectation for mocking the clear_background_color method

pub fn expect_window_focused(&mut self) -> &mut Expectation

Create an Expectation for mocking the window_focused method

pub fn expect_focus_window(&mut self) -> &mut Expectation

Create an Expectation for mocking the focus_window method

pub fn expect_audio_output_volume(&mut self) -> &mut Expectation

Create an Expectation for mocking the audio_output_volume method

pub fn expect_set_audio_output_volume(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_audio_output_volume method

pub fn expect_link_bind_device(&mut self) -> &mut Expectation

Create an Expectation for mocking the link_bind_device method

pub fn expect_set_link_bind_device(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_link_bind_device method

pub fn expect_inhibit_hotkeys(&mut self) -> &mut Expectation

Create an Expectation for mocking the inhibit_hotkeys method

pub fn expect_set_mouse_cursor(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_mouse_cursor method

pub fn expect_add_link_p2p_excluded_subnet(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_link_p2p_excluded_subnet method

pub fn expect_clear_link_p2p_excluded_subnets(&mut self) -> &mut Expectation

Create an Expectation for mocking the clear_link_p2p_excluded_subnets method

pub fn expect_link_time_since_last_received_packet_ns( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the link_time_since_last_received_packet_ns method