oden_plugin_rs::settings_api

Struct MockSettingsApi

pub struct MockSettingsApi { /* private fields */ }

Implementations

impl MockSettingsApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockSettingsApi

pub fn expect_read_int(&mut self) -> &mut Expectation

Create an Expectation for mocking the read_int method

pub fn expect_read_float(&mut self) -> &mut Expectation

Create an Expectation for mocking the read_float method

pub fn expect_read_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the read_string method

pub fn expect_read_bool(&mut self) -> &mut Expectation

Create an Expectation for mocking the read_bool method

pub fn expect_write_int(&mut self) -> &mut Expectation

Create an Expectation for mocking the write_int method

pub fn expect_write_float(&mut self) -> &mut Expectation

Create an Expectation for mocking the write_float method

pub fn expect_write_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the write_string method

pub fn expect_write_bool(&mut self) -> &mut Expectation

Create an Expectation for mocking the write_bool method

pub fn expect_project_path(&mut self) -> &mut Expectation

Create an Expectation for mocking the project_path method

pub fn expect_project_plugin_path(&mut self) -> &mut Expectation

Create an Expectation for mocking the project_plugin_path method

pub fn expect_application_type(&mut self) -> &mut Expectation

Create an Expectation for mocking the application_type method

pub fn expect_set_has_crashed(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_has_crashed method

pub fn expect_set_ui_disable(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_ui_disable method

pub fn expect_read_scene_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the read_scene_string method

pub fn expect_write_scene_string(&mut self) -> &mut Expectation

Create an Expectation for mocking the write_scene_string method

pub fn expect_exit_application(&mut self) -> &mut Expectation

Create an Expectation for mocking the exit_application method

pub fn expect_project_file_path(&mut self) -> &mut Expectation

Create an Expectation for mocking the project_file_path method

pub fn expect_publish_data(&mut self) -> &mut Expectation

Create an Expectation for mocking the publish_data method

pub fn expect_query_data(&mut self) -> &mut Expectation

Create an Expectation for mocking the query_data method

pub fn expect_plugin_param(&mut self) -> &mut Expectation

Create an Expectation for mocking the plugin_param method

pub fn expect_oden_version(&mut self) -> &mut Expectation

Create an Expectation for mocking the oden_version method

pub fn expect_license_hash(&mut self) -> &mut Expectation

Create an Expectation for mocking the license_hash method

pub fn expect_instance_uuid(&mut self) -> &mut Expectation

Create an Expectation for mocking the instance_uuid method