oden_plugin_rs

Struct InitParams

pub struct InitParams<'a> { /* private fields */ }

InitParams

OdenPlugin::init is called when the plugin is initialized. This is an appropriate place to create data for storing state and/or initializing threads.

InitParams is a struct that exposes the Oden Plugin API functions associated with the OdenPlugin::init function. The API is passed as a parameter to init(api: &InitParams) and can be accessed as shown in the example below:

Examples

fn init(api: &InitParams) -> Self {
    State { _dummy: false }
}

Implementations

impl InitParams<'_>

pub fn read_int(&self, key: &str) -> Option<i32>

Reads persistent i32 data from the key identifier in the .vproj file.

Examples

let key_value = api.read_int("key");

pub fn read_float(&self, key: &str) -> Option<f32>

Reads persistent f32 data from the key identifier in the .vproj file.

Examples

let key_value = api.read_float("key");

pub fn read_string(&self, key: &str) -> Option<String>

Reads persistent String data from the key identifier in the .vproj file.

Examples

let key_value = api.read_string("key");

pub fn read_bool(&self, key: &str) -> Option<bool>

Shortcut to read ints as bools

pub fn write_int(&self, key: &str, value: i32)

Writes persistent i32 data to the key identifier in the .vproj file.

Examples

api.write_int("key name", 1);

pub fn write_float(&self, key: &str, value: f32)

Writes persistent f32 data to the key identifier in the .vproj file.

Examples

api.write_float("key name", 1.0);

pub fn write_string(&self, key: &str, value: &str)

Writes persistent String data to the key identifier in the .vproj file.

Examples

api.write_string("key name", "string_value");

pub fn write_bool(&self, key: &str, value: bool)

Writes persistent bool data as an i32 to the key identifier in the .vproj file.

Examples

api.write_bool("key name", true);

pub fn project_path(&self) -> Option<String>

Returns the current project path, or None if the project is not saved.

Examples

let plugin_path = api.project_path();

pub fn project_plugin_path(&self) -> Option<String>

Returns the current project plugin path, or None if the path is unspecified.

Examples

let plugin_path = api.project_plugin_path();

pub fn application_type(&self) -> ApplicationType

Returns the current ApplicationType. Can be used to differentiate between Player, Producer and Streamer.

Examples

let application_type = api.application_type();

pub fn set_has_crashed(&self)

Signals to Oden that the plugin has crashed. This is used by the oden plugin panic handler

Examples

api.set_has_crashed();

pub fn set_ui_disable(&self, disable: bool)

Disable the Ctrl+H GUI in Oden.

If the disable arg is true than Ctrl+H functionality will be disabled.

Examples

 api.set_ui_disable(true);

pub fn read_scene_string(&self, key: &str) -> Option<String>

Reads String data from the key identifier in the scene data. The key identifier can be written as a string such as “key_value” and the value will read from the current scene. To read data from a specific scene use the following syntax “Scene Name@key_value”. For global plugins the latter alternative is required.

Examples

let key_value = api.read_scene_string("Scene1@key");

pub fn write_scene_string(&self, key: &str, value: &str) -> bool

Write persistent String data to the key identifier in the scene data. The key identifier can be written as a string as “key” and the value will be added to the current scene. To write data to a specific scene use the following syntax “Scene Name@key”. For global plugins the latter alternative is required.

Examples

api.write_scene_string("Scene@key name", "string_value");

pub fn exit_application(&self, force_exit: bool)

Will make the application gracefully exit without saving. With force_exit = false it will take into regard the “No Promt On Exit” setting and give a pop up to the user. If force_exit = true it will disregard the “No Promt On Exit” setting and exit on the next frame.

Examples

    api.exit_application(false);

pub fn project_file_path(&self) -> Option<String>

Returns the current project plugin file path, or None if the path is unspecified.

Examples

let plugin_file_path = api.project_file_path();

pub fn publish_data(&self, data_id: &str, data: &[u8])

Publishes data as data_id that is shared between plugins.

Oden offers a way to share data between plugins using named memory regions. One plugin can publish a named shared data and multiple plugins may query it.

pub fn query_data( &self, data_id: &str, expected_data_size: i32, ) -> Result<(Vec<u8>, QueryMetadata), QueryError>

Returns shared plugin data, metadata, and possible errors, or crate::QueryError if an error occurred.

Query data that have been shared between plugins. Oden offers a way to share data between plugins using named memory regions. One plugin can publish a named shared data and multiple plugins may query it. Supplying expected_data_size == -1 will return all data until a terminating byte’\0’

Also used for command line data: --plugin-data [key] [val]

pub fn plugin_param(&self, key: &str) -> Option<String>

Retrieve a plugin parameter

Plugin parameters can be passed on command line as [application] --plugin-param [key] [value] This function takes [key] as input and returns [value]

Examples

let window_title = api.plugin_param("window_title");

pub fn oden_version(&self) -> String

Returns the current Oden version, e.g. “1.23.4”

Examples

let version = api.oden_version();

pub fn license_hash(&self) -> Option<String>

Fetches the hash of the currently activated oden license

Example:

let _license_hash = api.license_hash();

pub fn instance_uuid(&self) -> Option<Uuid>

Returns the current plugin instance uuid.

Example:

let instance_uuid = api.instance_uuid();

impl InitParams<'_>

pub fn read_struct<T: Default + DeserializeOwned + 'static>( &self, key: &str, ) -> T

Reads an arbitrary deserializable struct from the key identifier in the .vproj file.

pub fn try_read_struct<T: DeserializeOwned + 'static>( &self, key: &str, ) -> Option<T>

See $crate::settings_api::SettingsExtApi::try_read_struct

pub fn write_struct<T: Serialize + 'static>(&self, key: &str, value: &T)

Writes an arbitrary serializable struct to the key identifier in the .vproj file.

pub fn modify_struct<T, F>(&self, key: &str, f: F)
where T: Default + Serialize + DeserializeOwned + 'static, F: FnOnce(&mut T),

Convenient way of working with settings-backed structs

pub fn read_scene_struct<T: Default + DeserializeOwned + 'static>( &self, key: &str, ) -> T

Reads an arbitrary struct that have been written to the scene data. The key identifier can be written as a string as “key_value” and the value will be added to the current scene. To write data to a specific scene use the following syntax “Scene Name@key_value”. For global plugins the later alternative is a must.

pub fn write_scene_struct<T: Serialize + 'static>( &self, key: &str, value: &T, ) -> bool

Write an arbitrary struct to the scene data. The key identifier is a string formatted as “key_value”. To read data from specific scene use the following syntax “Scene Name@key_value”. For global plugins the later alternative is a must.

pub fn modify_scene_struct<T, F>(&self, key: &str, f: F) -> bool
where T: Default + Serialize + DeserializeOwned + 'static, F: FnOnce(&mut T),

Convenient function to modify a stored struct in the scene data. The key identifier is a string formatted as “key_value”. To modify data from specific scene use the following syntax “Scene Name@key_value”. For global plugins the later alternative is a must.

pub fn publish_var_size_data<T>(&self, id: &str, data: &T)
where T: Serialize,

Publishes data as data_id that is shared between plugins. This is a utility function for publish_data to publish variable sized data.

Oden offers a way to share data between plugins using named memory regions. One plugin can publish a named shared data and multiple plugins may query it.

pub fn query_var_size_data<'a, T>( &self, id: &str, ) -> Result<(T, QueryMetadata), QueryError>
where T: for<'de> Deserialize<'de>,

Query data that have been shared between plugins. Returns shared plugin data, metadata, and possible errors, or crate::QueryError if an error occurred. This is a utility function for query_data to query variable sized data.

Oden offers a way to share data between plugins using named memory regions. One plugin can publish a named shared data and multiple plugins may query it.

Also used for command line data: --plugin-data [key] [val]

impl InitParams<'_>

pub fn upload_image( &self, id: i32, image_width: i32, image_height: i32, image: &[u8], )

Uploads an image resource with dimensions image_width x image_height.

The image MUST be in RGBA color format. The supplied id MUST be unique for every image.

This only needs to be done once for every image. The image can then be used efficiently every frame by referring to the id.

pub fn upload_image_ex( &self, id: i32, image_width: i32, image_height: i32, image: &[u8], params: &UploadImageParams, )

Uploads an image resource with dimensions image_width x image_height.

The supplied id MUST be unique for every image. Using an existing id will overwrite the previously uploaded image.

This only needs to be done once for every image. The image can then be used efficiently every frame by referring to the id.

Use params to specify the texture format

pub fn upload_font( &self, id: i32, glyph_size: f32, font_descriptors: &[FontDescriptor<'_>], )

Works similarly to upload_image but for TTF font files. Contact Voysys if support is needed for font functionality.

pub fn upload_shader(&self, id: i32, code: &str) -> bool

Uploads a shader for use with im draw

The supplied code must be glsl and declare a function with the following prototype: vec4 frag(Params p) { … }

where Params is

struct Params { vec4 tint; vec4 vertex_color; vec4 texture_color; vec4 back_color; vec3 pos; vec2 uv; float time; };

pub fn register_gl_tex_id(&self, image_id: i32, texture_id: u32)

Registers an OpenGL texture_id to an Oden Plugin image_id.

The image_id needs to be maintained by the plugin and MUST be unique for each image.

pub fn register_iosurface_id(&self, image_id: i32, iosurface_id: u32)

Registers a macOS IOSurface ID to an Oden Plugin image_id.

pub fn register_image_callback<F>(&self, stream_id: &str, callback: F)
where F: FnMut(VideoFrame) + Send + 'static,

Registers a callback for when Oden receives a non coded image from a camera

stream_id is a name of video source and the video stream id “name_of_the_entity:index_of_the_stream”

Examples

api.register_image_callback("test:1", |frame| {
    info!("{} {}", frame.size[0], frame.format);
});

pub fn set_window_title(&self, title: &str)

Sets the window title (overriding the default “Application Name - loaded_project.vproj”

The supplied title will override the default window title. Pass an empty string to clear the override and restore the default title.

impl InitParams<'_>

pub fn gl_loader(&self) -> GlLoader

Returns the OpenGL function loader that Oden uses.