oden_plugin_rs

Trait OdenPlugin

pub trait OdenPlugin {
    // Required methods
    fn init(api: &InitParams<'_>) -> Self;
    fn shutdown(self, api: &ShutdownParams<'_>);
    fn update(&mut self, api: &UpdateParams<'_>);
    fn gui(&mut self, api: &GuiParams<'_>);

    // Provided methods
    fn draw(&mut self, _api: &DrawParams<'_>) { ... }
    fn init_logger() -> bool { ... }
}

OdenPlugin Trait

Implement this trait for a type in order to register that type as a plugin entity type.

Required Methods

fn init(api: &InitParams<'_>) -> Self

Initializes a new instance of the plugin.

fn shutdown(self, api: &ShutdownParams<'_>)

Destructor for cleaning up resources when plugin instance is unloaded.

fn update(&mut self, api: &UpdateParams<'_>)

Called on each frame update during the plugin instances lifetime.

fn gui(&mut self, api: &GuiParams<'_>)

Called on each frame when the Oden config GUI is displayed. Function is responsible for displaying plugin configuration elements.

Provided Methods

fn draw(&mut self, _api: &DrawParams<'_>)

Called for each surface that oden want the plugin to draw to

fn init_logger() -> bool

Called to determine if the plugin wants the oden logger to be initialized

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.