oden_plugin_rs::update_api

Struct MockUpdateApi

pub struct MockUpdateApi { /* private fields */ }

Implementations

impl MockUpdateApi

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockUpdateApi

pub fn expect_dt(&mut self) -> &mut Expectation

Create an Expectation for mocking the dt method

pub fn expect_world_matrix(&mut self) -> &mut Expectation

Create an Expectation for mocking the world_matrix method

pub fn expect_camera_forward_ray(&mut self) -> &mut Expectation

Create an Expectation for mocking the camera_forward_ray method

pub fn expect_mouse_pos(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_pos method

pub fn expect_mouse_pressed_since_last_swap(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_pressed_since_last_swap method

pub fn expect_mouse_released_since_last_swap(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_released_since_last_swap method

pub fn expect_download_virtual_camera_image( &mut self, ) -> &mut Expectation<'static>

Create an Expectation for mocking the download_virtual_camera_image method

pub fn expect_push_recording_bytes(&mut self) -> &mut Expectation

Create an Expectation for mocking the push_recording_bytes method

pub fn expect_push_recording_data<T: AsBytes + 'static>( &mut self, ) -> &mut Expectation<T>

Create an Expectation for mocking the push_recording_data method

pub fn expect_recording_playback_channel(&mut self) -> &mut Expectation

Create an Expectation for mocking the recording_playback_channel method

pub fn expect_reset_view(&mut self) -> &mut Expectation

Create an Expectation for mocking the reset_view method

pub fn expect_set_fps_camera_angles_and_position(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_fps_camera_angles_and_position method

pub fn expect_set_fps_camera_fov(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_fps_camera_fov method

pub fn expect_set_entity_transform(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_entity_transform method

pub fn expect_mouse_ray(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_ray method

pub fn expect_viewport_info(&mut self) -> &mut Expectation

Create an Expectation for mocking the viewport_info method

pub fn expect_mouse_pos_px_with_offset(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_pos_px_with_offset method

pub fn expect_mouse_clicked_last_element(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_clicked_last_element method

pub fn expect_mouse_pressed_over_last_element(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_pressed_over_last_element method

pub fn expect_mouse_over_last_element(&mut self) -> &mut Expectation

Create an Expectation for mocking the mouse_over_last_element method

pub fn expect_virtual_camera_pixel_to_ray(&mut self) -> &mut Expectation

Create an Expectation for mocking the virtual_camera_pixel_to_ray method

pub fn expect_screen_space(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the screen_space method

pub fn expect_remove_last_element(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_last_element method

pub fn expect_monitor_info(&mut self) -> &mut Expectation

Create an Expectation for mocking the monitor_info method

pub fn expect_add_neural_network(&mut self) -> &mut Expectation

Create an Expectation for mocking the add_neural_network method

pub fn expect_remove_neural_network(&mut self) -> &mut Expectation

Create an Expectation for mocking the remove_neural_network method

pub fn expect_set_neural_network_model(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_neural_network_model method

pub fn expect_set_neural_network_input_video_stream( &mut self, ) -> &mut Expectation

Create an Expectation for mocking the set_neural_network_input_video_stream method

pub fn expect_neural_network_tensor_data(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the neural_network_tensor_data method

pub fn expect_trigger_neural_network_run(&mut self) -> &mut Expectation

Create an Expectation for mocking the trigger_neural_network_run method

pub fn expect_neural_network_info(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the neural_network_info method

pub fn expect_inference_settings(&mut self) -> &mut Expectation<'static>

Create an Expectation for mocking the inference_settings method

pub fn expect_set_neural_network_dla_core(&mut self) -> &mut Expectation

Create an Expectation for mocking the set_neural_network_dla_core method

pub fn expect_fps_camera_state(&mut self) -> &mut Expectation

Create an Expectation for mocking the fps_camera_state method

pub fn expect_prevent_close(&mut self) -> &mut Expectation

Create an Expectation for mocking the prevent_close method

pub fn expect_close_was_denied(&mut self) -> &mut Expectation

Create an Expectation for mocking the close_was_denied method