oden_plugin_rs

Module input_api

Input API

API for handling input from joysticks and gamepads.

Structs

Gamepad

A gamepad with named buttons and axes.

Enums

PowerState

Power state of the system.

Functions

add_gamepad_mapping

Add a gamepad mapping.

gamepad

Get the state of the gamepad if exactly one is connected.

gamepad_state

Get the state of a gamepad with the given id.

get_power_info

Get the power info of the system.

is_joystick_present

Check if a joystick with the given id is present.

joystick_state

Get the state of a joystick with the given id.